Python/requirements.txt
numpy>=1.24.0
scipy>=1.10.0
matplotlib>=3.7.0
pillow>=9.5.0
scikit-learn>=1.2.0
torch>=2.0.0
torchvision>=0.15.0
pandas>=2.0.0
h5py>=3.8.0  # For working with .mat files 